Boost multiprecision

+++ src/mesh/postscript_io.C
@@ -133,9 +133,9 @@ void PostscriptIO::write (const std::string & fname)
         // lower left corner
            << "0 0 "
         // upper right corner
-           << static_cast<unsigned int>( rint((x_max-x_min) * _scale ))
+           << static_cast<unsigned int>( rint(double((x_max-x_min) * _scale )))
            << ' '
-           << static_cast<unsigned int>( rint((y_max-y_min) * _scale ))
+           << static_cast<unsigned int>( rint(double((y_max-y_min) * _scale )))
            << '\n';
 
       // define some abbreviations to keep
